Common Methods separated
