+++ CI-API/Program.cs
@@ -64,6 +64,8 @@ builder.Services.AddScoped<IAdminService, AdminService>();
 builder.Services.AddScoped<IAdminRepository, AdminRepository>();
 builder.Services.AddScoped<IUserProfileRepository, UserProfileRepository>();
 builder.Services.AddScoped<IUserProfileService, UserProfileService>();
+builder.Services.AddScoped<ICommonRepository, CommonRepository>();
+builder.Services.AddScoped<ICommonService, CommonService>();
 builder.Services.AddScoped<ILandingPageRepository, LandingPageRepository>();
 builder.Services.AddAuthentication(x =>
 {
